Pruebas

+++ src/leer_archivo.cpp
@@ -1,4 +1,4 @@
-#include "leer_archivo.h"
+#include "include/leer_archivo.h"
 
 void leer_archivo (const char *archivo,std::vector<std::vector<int> >&param,int &camiones,double &capacidad_camiones,int &traileres, double &capacidad_traileres, int &clientes){
     std::ifstream ifs(archivo, std::ios::in);
@@ -10,7 +10,7 @@ void leer_archivo (const char *archivo,std::vector<std::vector<int> >&param,int 
     {
         std::stringstream ss(linea);
 	        std::vector<int> datos_cliente;
-	        double tipo_cliente;
+	        int tipo_cliente;
 	        while (ss >> tipo_cliente){
 	      	    datos_cliente.push_back(tipo_cliente);
 	        }
